Move snapshot finding functions to new file
